+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/quotes/cache/QuoteCacheService.kt
@@ -1,25 +1,20 @@
 package arrow.meta.ide.plugins.quotes.cache
 
 import arrow.meta.ide.dsl.application.cache.Id
-import arrow.meta.ide.dsl.application.cache.PersistentService
-import com.intellij.openapi.components.State
-import com.intellij.openapi.components.Storage
-import com.intellij.openapi.components.StoragePathMacros
+import arrow.meta.ide.dsl.application.cache.IdService
 import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.psi.KtFile
 import java.util.concurrent.ConcurrentHashMap
 
 /**
- * a QuoteCache implementation as a PersistentService, which stores Quote transformations on disk.
+ * a QuoteCache implementation as a IdService, which stores Quote transformations on disk.
  * This is an initial implementation, as it does not populate the index of other connected projects,
  * and the cache stored on disk get out of sync with external changes to the disk, projects sharing source files or other situations.
- * fixme must not be used in production, because caching PsiElement this way is bad
  * fixme cache both per module? modules may define different ktFiles for the same package fqName
  * TODO: register this via Meta.
  */
-@State(name = "QuoteCache", storages = [Storage(StoragePathMacros.CACHE_FILE)])
-private object QuoteCacheService : PersistentService<ConcurrentHashMap<KtFile, QuoteInfo>>, QuoteCache {
+private object QuoteCacheService : IdService<ConcurrentHashMap<KtFile, QuoteInfo>>, QuoteCache {
   override var value: Id<ConcurrentHashMap<KtFile, QuoteInfo>> =
     Id.just(ConcurrentHashMap<KtFile, QuoteInfo>())
